requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tensorflow==2.20.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
pillow==10.4.0
opencv-python==4.10.0.84
python-multipart==0.0.9
matplotlib==3.9.2
